adjust schema

+++ database/migrations/2018_08_22_081132_create_shoppingcart_table.php
@@ -11,20 +11,24 @@ class CreateShoppingcartTable extends Migration
      */
     public function up()
     {
-        Schema::create(config('cart.database.table'), function (Blueprint $table) {
-            $table->string('identifier');
-            $table->string('instance');
-            $table->longText('content');
-            $table->nullableTimestamps();
+        if (!Schema::hasTable(config('cart.database.table'))) {
+            Schema::create(config('cart.database.table'), function (Blueprint $table) {
+                $table->string('identifier');
+                $table->string('instance');
+                $table->longText('content');
+                $table->nullableTimestamps();
 
-            $table->primary(['identifier', 'instance']);
-        });
+                $table->primary(['identifier', 'instance']);
+            });
+        }
     }
     /**
      * Reverse the migrations.
      */
     public function down()
     {
-        Schema::drop(config('cart.database.table'));
+        if (Schema::hasTable(config('cart.database.table'))) {
+            Schema::drop(config('cart.database.table'));
+        }
     }
 }
